+++ package/CHANGELOG.md
@@ -1,5 +1,10 @@
 # TitleScreenQuickAccess
 
+## 1.0.2
+
+- Fixed tooltips stuck on screen when buttons are clicked with Enter key
+- Only show plugin config and custom level buttons when they're available
+
 ## 1.0.1
 
 - Added tooltips